Pandas Assignment 1(Heroes of Pymoli)
Assignment with ipynb file along with the pdf of codes. Having trouble running codes as it does not let me write and save codes in assignment folder. If i move the file to desktop i can write and save the code but cannot successfully run as it cannot recognize the source data.  I also lost my next assignmnet  (pycityschools) while troubleshooting this issue.
So am strating with the new file for pycityschools again.