API Homework
The homework file shows the approch to finish the homework. The work could not be completed as my API calls exceeded due to exceeded limit. I have emailed open weather map regarding the same but my key is not updated yet. So i did not get the data to work with but i do have the approach listed. I will complete the project asa i get the API key.