Final WeatherPy Homework File
I used a differnet API key today and got it done.  Please ignore weathertry as it does not have running codes, but the approach only.
WeatherSmile is the final file. I am able to save to picture and csv to my desktop only. Its not saving in the assignment folder . However code to save in both the folders are included.